+++ config/routes.rb
@@ -318,6 +318,8 @@ Rails.application.routes.draw do
         post :add_question
         post :report_lookup
         get :search
+        get :checks_search
+        get :values_search
       end
     end
   end