+++ CHANGELOG.md
@@ -19,6 +19,10 @@
   - Calculated variables are now formatted as numeric in SAS exports
   - `Sheet Created` is now included in exports
   - Missing sheets now appear in exports and are labeled in SAS and R exports
+- **Sheet Changes**
+  - Improved the autocomplete capability for search filters
+    - Check names will be recommended for autocompletion
+    - Variable values will be recommended for autocompletion
 
 ### Bug Fix
 - Fixed a bug in computing the correct column in the R export script